requirements/base.txt
django==1.7.1
pygments==2.0.2
django-pygments==0.1
markdown==2.6.2
django-markdown==0.8.4
django-haystack==2.4.0
Whoosh==2.7.0
jieba==0.37
django-adminfiles==1.0.1
sorl-thumbnail==12.2
pillow==2.8.1
django-cleanup==0.3.1
django-templateaddons==0.1
mysql==0.0.2
MySQL-python==1.2.3
